swipe eliminar mediante id

+++ app/src/main/java/com/example/pachanga_joseasanchezlopez/ui/home/HomeFragment.java
@@ -61,7 +61,7 @@ public class HomeFragment extends Fragment {
 
         eventos = new ArrayList<>();
         rvEventos = root.findViewById(R.id.rvEventos);
-        refreshLayout=root.findViewById(R.id.swipeRefresh);
+        refreshLayout = root.findViewById(R.id.swipeRefresh);
         rvEventos.setLayoutManager(new LinearLayoutManager(getContext()));
 
         llenarLista();
@@ -82,7 +82,7 @@ public class HomeFragment extends Fragment {
             }
         });
 
-        ItemTouchHelper.SimpleCallback simpleCallback=new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.RIGHT) {
+        ItemTouchHelper.SimpleCallback simpleCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.RIGHT) {
             @Override
             public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
                 return false;
@@ -90,22 +90,26 @@ public class HomeFragment extends Fragment {
 
             @Override
             public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-                FirebaseFirestore db=FirebaseFirestore.getInstance();
-
-                String evento=eventos.get(viewHolder.getAdapterPosition()).getId();
-                System.out.println(eventos.get(viewHolder.getAdapterPosition()).getNombre());
-                db.collection("pachangas").document(evento).delete();
-                eventos.remove(viewHolder.getAdapterPosition());
-                adapter.notifyDataSetChanged();
+                FirebaseFirestore db = FirebaseFirestore.getInstance();
+                if (eventos.get(viewHolder.getAdapterPosition()).getCreador().equals(auth.getCurrentUser().getEmail())) {
+                    String evento = eventos.get(viewHolder.getAdapterPosition()).getId();
+                    System.out.println(eventos.get(viewHolder.getAdapterPosition()).getNombre());
+                    db.collection("pachangas").document(evento).delete();
+                    eventos.remove(viewHolder.getAdapterPosition());
+                    adapter.notifyDataSetChanged();
+                }else{
+                    return;
+                }
             }
         };
 
-        ItemTouchHelper itemTouchHelper=new ItemTouchHelper(simpleCallback);
+        ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleCallback);
         itemTouchHelper.attachToRecyclerView(rvEventos);
 
         return root;
     }
-    ArrayList<NuevoEvento> recogidos=new ArrayList<>();
+
+    ArrayList<NuevoEvento> recogidos = new ArrayList<>();
 
     private void llenarLista() {
         eventos.add(new NuevoEvento("25/09/2022", "Mi casa", "Avengers", "21:30", true, "marina@gmail.com"));
@@ -118,8 +122,8 @@ public class HomeFragment extends Fragment {
         mDatabase.collection("pachangas").get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
             @Override
             public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                for (QueryDocumentSnapshot t:task.getResult()) {
-                    recogidos.add(new NuevoEvento(""+t.get("id"),""+t.get("fecha"),""+t.get("lugar"), ""+t.get("nombre"), ""+t.get("hora"), t.getBoolean("privado"), ""+t.get("creador")));
+                for (QueryDocumentSnapshot t : task.getResult()) {
+                    recogidos.add(new NuevoEvento("" + t.get("id"), "" + t.get("fecha"), "" + t.get("lugar"), "" + t.get("nombre"), "" + t.get("hora"), t.getBoolean("privado"), "" + t.get("creador")));
                 }
             }
         });
